feat: Sprint 4.1 — 真机部署闭环（Mac/Windows/香橙派三节点）+ scope 路由
entry 模型修复（真机委派实测发现）：
- 容忍推理模型输出任务 JSON 前的 chain-of-thought preamble（jsonLeadInMax 100→2000），
  同时用「示例/例如/仅供参考」标记守住"散文里的示例 JSON 应回退为 answer"
- decodeEnvelope 区分语法错误（回退 answer）与 schema 类型错误（上报，不再静默降级）
- ResourceProfile.ram_gb/gpu_vram_gb int→float64（模型会输出 0.1 这类小数）
- DeepSeek Anthropic 端点兼容：去掉 string 形式 tool_choice、点号工具名改 memory_add 等

路由：尊重用户点名节点
- scheduler.Route 新增 preferred：spec.scope 命名的节点在线且匹配时优先于调度等级
- PreferredNode 经 TaskInput / TaskDelegatePayload 随链传递，子调度器多跳转发同样生效
- 修复前「在香橙派上运行 uname -a」会因 sys:info 共享能力被 Win 顶掉，现按点名路由

部署：
- 新增 deploy/windows/（SSH 密钥安装脚本 + 启动脚本），Windows 节点开机自启计划任务
- 香橙派 shared_secret 经 systemd EnvironmentFile 注入（data/network-secret，gitignored）
- 三节点全互联 peers（香橙派补上 Windows 对端）

+++ internal/entry/types.go
@@ -63,8 +63,8 @@ type TaskSpecDetail struct {
 // ResourceProfile is a coarse resource hint, not a safety rating (design doc
 // §7.3; complexity/risk are recorded but do not auto-switch model tiers in MVP).
 type ResourceProfile struct {
-	CPU          int    `json:"cpu"`
-	RAMGB        int    `json:"ram_gb"`
-	GPUVRAMGB    int    `json:"gpu_vram_gb"`
-	DurationHint string `json:"duration_hint"` // short | long
+	CPU          int     `json:"cpu"`
+	RAMGB        float64 `json:"ram_gb"`
+	GPUVRAMGB    float64 `json:"gpu_vram_gb"`
+	DurationHint string  `json:"duration_hint"` // short | long
 }
